Renamed controllers/index.js to handleUpload.js, moved multer import in server/index.js to more appropriate place
